Added Login and Register page with role-based redirection
If the user is new he will register and select his role only once and then the user's data will be stored in the database. Later when the user visits again he do not need to register again he will simply go to the login page and enter his email and password and the data will be fetched from the database and the user will be redirected to their desired dashboard based on their role.